Remove most uses of usize/isize

+++ src/pretty/lib.rs
@@ -61,7 +61,7 @@ impl Doc {
     }
 
     #[inline]
-    pub fn nest(self, off: usize) -> Doc {
+    pub fn nest(self, off: u64) -> Doc {
         let Doc(doc) = self;
         Doc(Nest(off, Box::new(doc)))
     }
@@ -72,7 +72,7 @@ impl Doc {
     }
 
     #[inline]
-    pub fn render<W: io::Writer>(&self, width: usize, out: &mut W) -> io::IoResult<()> {
+    pub fn render<W: io::Writer>(&self, width: u64, out: &mut W) -> io::IoResult<()> {
         let &Doc(ref doc) = self;
         best(doc, width, out).and_then(|()| out.write_line(""))
     }
